tests/requirements.txt
pytest==8.2.0
azure-storage-blob==12.19.0
pika==1.3.2
requests==2.32.3
docker==6.1.3 